fix(driver,unix): remove redundant cast in NsSockAccept and improve macOS signal handling
- Removed unnecessary cast to int in the NsSockAccept() return
  statement, since SockAccept() already returns an int-compatible
  SockState value.  This avoids redundant type conversion and improves
  type clarity.

- Added macOS-specific handling in unix.c to undefine sigaddset (and
  similar macros if expanded) so that the real function
  implementations are used.  This prevents compiler warnings such as
  "-Wsign-conversion" caused by the macro-based bit manipulation
  variants on macOS SDKs.

+++ nsd/driver.c
@@ -3705,7 +3705,7 @@ SockAccept(Driver *drvPtr, NS_SOCKET sock, Sock **sockPtrPtr, const Ns_Time *now
 int
 NsSockAccept(Ns_Driver *drvPtr, NS_SOCKET sock, Ns_Sock **sockPtrPtr, const Ns_Time *nowPtr, void *arg)
 {
-    return (int)SockAccept((Driver *)drvPtr, sock, (Sock**)sockPtrPtr, nowPtr, arg);
+    return SockAccept((Driver *)drvPtr, sock, (Sock**)sockPtrPtr, nowPtr, arg);
 }
